(CPR-248) Add link to files list in link DSL method
Previously links were not added to the files list which meant that they
would not be collected if they fell outside of the directories being
watched by vanagon. This commit adds them to the files list.
The cp calls in the spec and rules files need the -R flag added to
ensure that links are not dereferenced and are copied as is, which would
otherwise break packaging.