(CPR-250) Ensure directories are created after initial file list
Previously the file-list-before-build could happen after the base
directories were created, which had the side effect of making it appear
as though those directories were not created by the current project.
This commit addresses that issue by ensuring that the
file-list-before-build is generated before any directories, and as early
as possibly in the Makefile.