(maint) Always drop etc, usr and var in IPS manifest
etc, usr and var are delivered by the core-os package in solaris 11.
This means that other packages that try to include these directories
will often run into conflicts against that base package. It is a common
practice to drop those directories in IPS manifests. It is also safe if
those directories aren't already present.