(RE-5679) Add package_type to platform
We need to be able to conditionally build certain packages. This commit
adds in the package_type param to the platform object. With this change,
we can now, if desired, specify that package type for a given platform
in a given project. If no package_type is specified, then it does the
default behavior defined by the specific platform file being loaded.