(RE-5979) Solaris-11 VERSION file should match other platforms
On Solaris 11, the version transformation to an IPS-compatible
format (minus alpha characters) also makes it into the VERSION
file.

This updates the logic to operate on a duplicate object instead
of the actual version variable when making the transformation,
so that the transformed value only winds up where we explicitly
ask for it.