(maint) Allow customization of 'find' and 'sort' when processing files
On windows, we need to make sure we're using the cygwin provided
binaries for find and sort. Otherwise, we find ourselves in the
unfortunate situation where windows find and windows sort do not
actually do what we are expecting. As a consequence, the file lists we
create and use to populate the tarball and the package are not populated
correctly and we end up with an empty package. This is not a positive
thing.