(RE-5680) Teach vanagon how to build a nuget package
This commit adds in the necessary information to build a nuget package
on windows. The work in this PR will need to be modified once we have
the procedure for building multiple types of packages on a single
platform. Eventually, much of this work will be moved into
windows/nuget.rb, ideally. However, we're just not there yet.