(RE-5974) Add lock_manager gem to vanagon project
In preparation for creating a new engine for vanagon to use when
building on raw hardware devices, the lock_manager gem will be required
to work with the current node lock system. This system is currently in
use by enterprise-dist and kerminator.

The goal of lock_manager is to be a library that can be used by other
tools in Puppet Labs engineering services.