(RE-5974) Vanagon should lock hardware when building
In the current build environment, vanagon can just trample any other
build going on. This is bad. In the PE 2.y days, we built a crude
locking system for nodes. That system was improved a little bit to have
a kerminator (hubot) interface.

This commit expands upon that by having vanagon lock a hardware node
when the build starts and unlock it when it finishes. It also adds a new
platform dsl method around build_host. This allow for a hostname to be
specified in the platform definition instead of passing a build target
on the command line. This is how vanagon knows to use the hardware
engine rather than just using base.

DSL Example: (in a platform config definition)

  plat.build_host "pe-aix-71-builder.delivery.puppetlabs.net"

The engine will lock the node upon build invocation, using the
lock_manager gem. The engine will issue a polling_lock request (meaning,
it will wait until a lock is granted).  One the build completes or
fails, the lock is removed via the Hardware engine's teardown method.

The engine also outputs and logs when requesting lock and unlock operations.

For configuration, we need LOCK_MANAGER_HOST set, or else it defaults
to redis as a shortname. You can also set LOCK_MANAGER_PORT which uses
the normal redis port 6379 by default.

The user that will own the lock is $USER.

Example build: (trimmed)

    $ bundle exec build puppet-agent aix-7.1-ppc
    Loading libxml2
    ...
    Hardware Locked: pe-aix-71-builder.delivery.puppetlabs.net
    ...
    mkdir -p output/aix/7.1/PC1/ppc
    cp /var/tmp/tmp.vx9cWRlvoa/rpmbuild/*RPMS/**/*.rpm
    ./output/aix/7.1/PC1/ppc
    Removed hardware lock: pe-aix-71-builder.delivery.puppetlabs.net.

vanagon_hosts.log

    D, [2015-11-20T05:30:09.245834 #17013] DEBUG -- vanagon: Hardware engine invoked.
    I, [2015-11-20T05:30:09.252352 #17013]  INFO -- vanagon: Polling for a lock on pe-aix-53-agent.delivery.puppetlabs.net.
    I, [2015-11-20T05:30:09.254071 #17013]  INFO -- vanagon: Lock acquired on pe-aix-53-agent.delivery.puppetlabs.net.
    I, [2015-11-20T05:32:32.536856 #17013]  INFO -- vanagon: Removing lock on pe-aix-53-agent.delivery.puppetlabs.net.