(CPR-239) On Solaris 10, vanagon projects change group ownership from 'bin' to 'sys'
Before this commit vanvagon runs would cause the permissions of the /lib /lib/svc and /lib/svc/method directories to change from "bin" to "sys". This was due to incorrectly generating the solaris 10 prototype file with the pkgproto command. This commit will fix the issue by excluding /lib and /bin directories from the prototype file as we do not want to change anything with those.

NOTE**** Technically we are incorrectly generating the prototype file for solaris 10; as stated here: http://docs.oracle.com/cd/E19120-01/open.solaris/817-0406/ch2buildpkg-30163/index.html we should be generating a pkginfo file before running pkgproto. Vanagon works as it is now, and pkgproto is not complaining, so there seems to be no problem. However when installation occurs on the target machine for puppet-agent pkgadd is technically acting out of the root directory, which may not be ideal. Again this does not seem to cause any problems but it should be noted that technically there is another step that could be included.