(maint) Handle git versioning scheme in nuspec template
Nuget is challenging when it comes to versions. We can only have four
numeric elements separated by periods. We have the option of specifying
a build version, but this must be the last element in the version tag
and must be a dash followed immediately by a letter and then either
numbers or letters but no punctuation.

If this is a build from a tag, we use the tag plus the release number to
version the project. If it's bot from a final tag, we make sure to
append the short sha with a dash.

Gross.