(maint) Allow platforms to specity the command to copy files
In the Makefile, we're defaulting to 'cp' as the copy command. This
generally is fine, but it doesn't work for special environments like
windows. We have some situations where we need to isolated any cygwin
commands from the path, but we still need to invoke them. This lets us
define where to get the copy command from.