(CPR-269) Assume unknown extension types are just files
Prior to this patch, we had to explicitly add every extension we wanted
to use a source with vanagon. This resulted in lots of small patches
with a simple addition of .json or .txt or something. It also meant
that to use those extensions, a release of vanagon was needed. This
slows down work.

Instead, we'll default to any extension that is not explicitly called
out as an archive, is simply a flat file and able to be used that way.
The net result here is that only new archive extensions will require
patches.