(CPR-264) Rewrite chocolatey packaging scripts
Our current nuget packaging scripts make a few assumptions that have
turned out to be false. Namely, these assumptions surround the path
specifications for where we are installing. Previously, we were using a
list of files and directories picked up from the project and component
configs. This was fine, but we have a better source for this
information. We have a file-list.txt file that details all of the files
and directories (even if they are empty) that are to be included in the
package. These paths have already been normalized to deal with spaces
and shortcuts (if any are used). So any use of `Progra~1` will already
have been expanded to `Program Files`. This is especially important
because we cannot guarentee that `Progra~1` will expand to what we want
it to be on the system the package is installed on. We have to expand
all of these shortcuts prior to packaging anything up. We use these
powershell scripts to then copy over files and directories from the
chocolatey install location to the final install location on the system.

The uninstall script only cleans up files. It does leave an empty
directory system in place.