(maint) Normalize path shortcuts
On windows, we need the ability to deal with spaces in path names. To
handle spaces, we need to allow the use of MS-DOS shortcuts when
specifying directories (i.e., `Progra~1` expands to `Program Files` in
most cases). However, we do not want to include any references to these
shortcuts in our final file lists. In order to remove them, we need to
normalize the paths. This commit uses cygpath to take care of that
process. Cygpath will expand any shortcuts on the build system (which is
more reliable and predictable then on a system in the wild).

Super gross, I know. I didn't know what else to do. There are
enough assumptions on cygwin currently, I figured it was safe to add one
more.