(maint) Remove unneeded methods
Prior to this commit, we were using `/cygdrive/c` as our path prefix due
to a few different bugs in vanagon. Those bugs have been fixed, and we
can rely solely on `C:/` as our prefix in most cases. The only exception
that I have found so far is setting `PATH` in a cygwin environment. This
requires the use of `cygpath`. However, since this is set on a project
level, my thoughts are that we should handle this in each project.