Add upstream URL to component sources
When trying to work with puppet-agent and other PL software outside of
Puppet Labs, things are difficult. This is an attempt to make it
slightly easier. This commit adds a new DSL method called 'upstream_url'
to a component source. This URL is used for fetching content if the
'url' specified fails.

The advantage of this is that we'll now have the upstream URL in the
component, and you won't have to spend your first 25 minutes download
sources and making a mirror on a webserver to use vanagon's url rewrite
engine.

Basically, this just means you need something like:

    pkg.upstream_url "http://ftp.gnu.org/gnu/#{pkg.get_name}/#{pkg.get_name}-#{pkg.get_version}.tar.gz"

Upstream URL is always the second choice behind pkg.url.

This patch has no bearing on handling of git urls.