(maint) Move templates dir to resources
When building packages with vanagon, we need more then just templates.
There are occasions where we need static packaging resources in addition
to or instead of templates. This is all dependent on the type of package
we are creating. However, it's misleading to have static resources in a
directory called `templates`. This commit renames the templates
directory to resources. This directory will contain both static and
dynamic resources to be used when creating packages.