(CPR_272) (maint) refactor git() function
The git function under vanagon utilities takes
an arbitrary sized nuber of strings as it's
only parameter, this was updated so that
instead it takes two params: the first is
simply one string representing the command
to be executed in git. The second is a T/F
value indicating whether or not to raise an
error if the git command fails. This was
refactored to facilitate rspec failures when
some functions expected the non-zero return from
git as input