(maint) Hardware engine should release lock upon failure
Prior to this, the hardware engine attempted to release the lock it had
on a system when regardless of the outcome of the vanagon build.
However, due to an if statement that could never evaluate to true, that
wasn't happening.

In debugging the problem, there was no way to figure out which engine
was being invoked. This commit adds a @name attribute to all engines so
other parts of the codebase can know what engine is being invoked.

This now means that if the hardware engine is being used, the ensure
block in driver.rb will release the lock regardless of pass/fail on the
build.