(maint) Allow zip archives to be unpacked with 7za
On windows, we need to use 7za to unpack zip archives. We don't have
`unzip` available and `gunzip` throws an error. The only problem is we
don't have any guarantees 7za will be on the path. It is installed as a
part of chocolatey, but we need to set path in the component in order to
ensure we have access to it.