(RE-6455) Create directorylist.wxs
This commit will add functionality to use the existing
get_services functionality for a project to generate
a directorylist.wxs file that contains the wix structure
of directories containing binaries for the project.

Because more than one service can point to the same binary
the directorylist may contain more than one directory wix
element at the top of the directory tree. In WiX this is
fine as long as any elements that are supposed to point
to the same directory have different IDs and no "name"
attribute.