(RE-6455) Create directorylist.wxs
This commit will add functionality to use the existing
get_directories functionality for a project to generate
a directorylist.wxs file that contains the wix structure
of directories for the project.

As part of this feature this commit also extends the current
VANAGON::COMMON::PATHNAME structure to include "id" for
use in the wix structure. it does not specifically extend
the actual COMMON::PATHNAME class but instead creates a
new class to extend pathname under the namespace
VANAGON::WINDOWS:PATHNAME