(CPR-294) Add #clean_environment helper for #local_command
Currently `gem install` will fail if Bundler is in use,
due to Bundler setting specific "RUBY_" environment variables.
When a local command (or series of commands) is dispatched, the polluted
environment inherited by the new subshell causes files to be
written to the wrong location (assuming that they're written at all).

Adding a #clean_environment helper to Vanagon::Utilities and
wrapping #local_command with it should ensure that local commands
are dispatched without a Bundler-polluted environment.