(PE-13450) AIX upgrade leaves services stopped
Currently, due to the way RPM scriptlet ordering works, on AIX
we are stopping the service *after* we restart the service on upgrades.

This makes a couple of changes.

First, make the service stop in the %preun conditional on package
uninstall, like the rest of the platforms.

Second, moves the upgrade %post actions to %postun, as that's more likely where
we actually want those to happen.

Unfortunately, neither change will help users who already have a version
of P-A installed, as the preun/postun scripts are run from the *previous*/
already installed package during upgrades, but this should fix the problem
going forward.