(RE-6464) Add rspecs to test/spec functionality
Create rspec tests to demonstrate and validate the expected behaviour of
the Project Specific/Vanagon Generic Merge operation. This also requires
a set of supporting files under the spec/fixtures directory.

The purpose of these tests is to validate the file movements including
the renaming of .erb templates. It is assumed that the ERB translation
is tested seperately.

Some modifications to the windows.rb file were necessary to setup the
test directories rather than relying on the existing of a project
specific (e.g. puppet-agent) repo.