(RE-6640) Filter out service files from harvest
This updates the service file handling to for product specific wix
service file fragments. The service .exe still needs to be filtered out
from the harvest, so the list of service .exe's are placed in the filter
file.

The project.filter.xslt file has also been renamed to simply
filter.xslt.  This file is given a generic name as there may be a need
in the future to filter out other files from the harvest.

The nssm.exe file is not being correctly matched in the harvest - this
appears to be an issue with the build itself (the nssm and some other
files are not being dropped in the correct folder). This issue will be
checked, ticketed if necessary and handled in seperate PR.