(MAINT) Convert filter.wxs.erb file to Unix Mode
This fixes an irritant in the ERB translation which leads to multiple
blank lines in the filter file. While not affecting functionality, it
represented an untidyness that "made me sad" in the PDX vernacular.

Although it looks like all lines have changed, in fact all that was done
was to run the :set fileformat=unix in vi.