(RE-6476) rework directorylist and run heat.exe twice
This commit adds several new parts to the current vanagon
work.

First off, we now realize we need to run heat.exe twice.
and to go along with this we need to update the
directorylist work such that it creates INSTALLDIR
and APPDATADIR references for the two new heat runs
to use.

To facilitate the new functionality in the directorylist
work the following was changed:

* the generate_wix_dirs function was updated to take in
a generic list of items, not specifically services

* added a generate_wix_structure function above the
generate_wix_dirs function to create the generic items.

* directorylist now contains serveral explicit directories
pertaining to the install roots. These will never change
so instead of autogenerating them we just explicitly
state them in this file, this serves as some extra
documented work so the user knows what the install root
looks like

Several things have changed in the input flags to heat.exe
and candle.exe, comments have been added to those for clarity

As a part of this work, rspec has been updated in order to
actually run.