(RE-6743) Vanagon generic registry entries
These entries are being over-ridden by a more Puppet-Agent specific set
of registry entries for puppet-agent.

Rather than deleting this file, it is being left in place with comments
to give guidance on how to set registry entries.

This model will also be applied to the properties and other common
files.