(RE-6520) Specify arch when generating Debian pkgs
This allows cross-compiled packages to use correct tools (such as the
target arch strip command) when generating the package. When
project.noarch is set, this becomes a no-op.

Note: you cannot have a space between the -a and the arch, which
only works for the dpkg-buildpackage in Debian 8 and later.