(maint) remove the checksum for local sources
If a source file is local to the vanagon project and
not fetched from http, there is no reason to check the
sum for the file since it's already in a repo and local
so git will handle keeping the version current and any
conflicts between forks.

This commit removes the validate step for local files
by making the "verify" function just do nothing for
local files.