(maint)(RE-6995) remove the checksum for local sources
This commit adds a new Component::Source child called
LocalSource to handle local files. This functionality
exists in the http child class of source and this commit
removes the functionality from that class and moves it to
a new one. With the new class we can now avoid checksumming
local files by making "verify" a noop function.

The new class comes with spec tests and updated spec
tests for source