(maint) Set verbosity when driver is created
Previously, if verbose or preserve were specified on the command line,
the values would not be applied to the Driver object until after the
`initialize` method returned.

The verbose and preserve options are already passed to the initialize
method, so set the verbose/preserve values as soon as possible. This
way verbose/preserve are correct when loading the platform and project
DSL.