(maint) Ensure engine name matches engine class
Previously, calling `Engine#name` for docker, hardware, and local engines
returned 'base' instead of the engine-specific name, e.g. 'docker'. The
reason was because those engines set their name and then called `super`,
and the Base superclass sets the name back to 'base'.

This commit moves the 'super' call to the very beginning, so that the
subclassed engine always wins when it comes to setting instance
variables.

Also this commit adds a `name` reader for each engine that returns its
hardcoded name. There's no need to store the name as an instance
variable or provide an accessor for changing the name of an existing
engine.

Also the `local` engine appeared to allow a target to be specified, but
in reality that could never happen -- because if a target is specified,
we always create a `base` engine. So modify the local engine to
explicitly set the target to `local machine` as it always did before.