(QENG-3556) Add Driver#build_host_info method
Each vanagon project is responsible for mapping a build target to the
host on which the build should be performed. In most cases, we do native
compiles, so the build host and target are the same type of machine, but
the build and target names use different naming conventions, e.g.
vanagon builds the ubuntu-15.10-amd64 target on an ubuntu-1510-x86_64
vmpooler template.

In other cases, the build and target names are binary compatible, but
not the exactly same, e.g. we build the el-7-x86_64 target on a
centos-7-x86_64 vmpooler template.

And in other cases we cross compile, e.g. we build Solaris SPARC on
x86_64, Windows x86 on x86_64, eos-4-i386 on fedora-14-i386.

This commit adds a method to Vangon::Driver to return build host
information, so that we can request the correct build host outside of
vanagon, and once the resource becomes available, pass it to vanagon.

The build host info is engine-specific. The pooler engine uses the
required `vmpooler_template` attribute of the platformcomponent. The
docker engine uses the `docker_image` required attribute. The hardware
engine uses the first entry from `build_hosts`. Requesting the build
host type for any other engine (base or local) will raise an exception.