(maint) Update the directorylist creation for windows to use uniq id
It's posible to cause a failure in WiX by registering two services on paths
that are different but have the same named dir somewhere, for example,
registering two services under:

    /dir/bin

    /other_dir/bin

will cause a failure because when we create the directorylist.wxs file the
"bin" directories will both have the same "Id" field because we just used
name there. This commit appends values to the Id field that will ensure
uniqueness.

This commit upates both the windows platform object and rspec