(CPR-237) Update rubocop to 0.41.2
I had to make a few compromises in order to get both tests and rubocop
to pass. Unfortunately, a lot of the checks are set to fairly high
thresholds. Higher than I would prefer. This lets me get away with the
minimum amount of refactoring, but leaves us vulnerably to overly
complex methods, etc. in future code.