(RE-6200) Use Ruby-Git instead of homebrew Git functions
Ruby-Git might not be the One True Git Library that solves all of our needs
but it's generally better than abstracting out "%x()"-wrapped calls to Git
ourselves.

I've extended it slightly to support Git submodules and unshallowing repos
that were shallow-cloned but we're now using it basically as-is. In theory,
this will potentially make porting to a different Git library (like Rugged)
easier, since we've just got a handful of composed functions instead of an
entire implementation to port.