(MAINT) Collapse Source handling methods
The original plan was to have a "best guess or specific definition" for
initializing a new Source object. But the "best guess" functionality now
relies on the Source subclasses providing some sort of "is this valid?"
functionality that was not part of the original plan. With this new functionality
the "best guess or specific definition" plan is abandoned.