(VANAGON-8) Add the ability to only build single components in vanagon
The use case for devs is they need to only build a single component and it's
dependencies. This PR solves this by adding a filter_components function to
the project class that recursively compiles all the components build
dependencies (and the build dependencies' dependencies and so on) together in
to one array. The driver class will use that function and replace the project
objects' components array with the output of filter_components.

Vanagon is built well enough to handle the rest, all build steps and makefile
rules can remain the same.

This should be a strictly additive change, and should not affect any current
build processes.

Usage: bundle exec build <project> <platform> [--only-build <components>]
where <components> is a comma seperated list