(PA-569) Enable brp strip scripts for Cisco platforms
Our cisco-5/7 build platforms are missing the definition of
__os_install_post, so we have to manually set it here. This
then enables the brp strip scripts, which are needed to reduce
the size of our agent packages for these platforms. This is
necessary as some targets will fail to even install the package
without stripped binaries because yum's package cache is kept
on a partition with ~50MB of free space.