(QENG-4426) Add `always_be_scheduling` engine
We have an internal need for supporting the management of hardware build
hosts through a common centralized scheduler, specifically
[always-be-scheduling](https://github.com/puppetlabs/always-be-scheduling)
(ABS).

Prior to this work, vanagon would manage hardware itself: configs would
list the specific hardware resources which were known for building on a
given platform/target. We would allocate the hardware instance, use the
`lock_manager` to control access, etc.

In a world of scheduled hardware, allocation, lock management, setup,
teardown, etc., are all Someone Else's Responsibility™.  This change
allows authors to opt-in to an "always_be_scheduling" engine, which
leaves the host management up to a third-party scheduler for certain
platforms.

Additionally, as we wish
[vmpooler](https://github.com/puppetlabs/vmpooler)-based vanagon builds
in the always-be-scheduling world to allow ABS to manage assignment of
vmpooler resources, here we allow the `always_be_scheduling` vanagon
engine to manage those resources as well.

This is feature-flagged behind a command-line switch and can be enabled
like so:

``` shell
$ build_host_info puppet-agent aix-5.3-ppc
{"name":"pe-aix-53-builder.example.com","engine":"hardware"}

$ build_host_info puppet-agent aix-5.3-ppc --engine always_be_scheduling
{"name":"aix-53-ppc","engine":"always_be_scheduling"}
```

In project platform configuration files we introduce the
`abs_resource_name` attribute. When specified, the
`always_be_scheduling` engine will use that `abs_resource_name` value as
the returned platform name. If this is not available, the
`vmpooler_template` value will be used.  If neither is available, the
platform name string will be returned unchanged.