(MAINT) Add a Rakefile
We have a number of tasks around linting, testing, and building
Vanagon that Rake can automate for us. I know we're all burned out on
Rake but these are the sorts of problems it exists to help solve. We
avoided adding a Rakefile for fear of it accidentally becoming the
standard interface for Vanagon but it's time to rip the band-aid off.