(VANAGON-43) Initial pass at metrics gathering
This commit adds in the foundation to begin metrics gathering. As is,
when merged, it will not begin to gather metrics. However, once we have
the ability to set environment variables at the project and platform
levels, we will simply have to pass in a server and port in order to
begin metrics collection.

This is also only the first step. We need to extend this work to support
multiple metric collection methods. That work will all manifest in
modifications to the profiling_shell script.