(FEATURE) Add `render` command
`render` will generate just the Makefile for a given Vanagon project
and place it in a temporary directory, where it can be inspected
without interrupting/halting/aborting a project build. It will install
no dependencies, and do no engine/build target provisioning. This is
intended to help developers fine-tune any project changes that would
affect the Makefile before a project is built.