(MAINT) Loosen validation for Env. values
Make variables don't have the same restrictions as Shell variables.
This commit will loosen the requirements/validation around Env.
variables so that it's possible to define values that are Make-
appropriate. If this turns out to be a bad idea or we need additional
complexity to handle Make & Shell variables, we can iterate on that
later.