(VANAGON-40) Add Platform Env. DSL methods
The Platform DSL can now define values in its Environment, which will
be merged with the Project Environment and rendered into the top scope
for the final Makefile. There's also a handful of Platform-specific
env. vars. that we now get "for free", like VANAGON_PLATFORM_NAME.

Additionally, some method variables in the Platform DSL have been
renam ed, because `name` is a reserved word and we ran the risk of
weird collisions if we aren't careful. This is something we should
probably try to clean up when we see it.

Finally, Makefile.erb has been updated so that it will now use a
merged Environment in the top scope.