(VANAGON-40) Component Env. Vars, part 1
Components already support a crude form of env. var. management, using
a series of clunky hashes. We should try to preserve the interface for
defining those env. vars. while migrating to the new conceptual model
for Makefile env. vars. This commit updates the existing spec tests to
make sure that they're returning the right values, and starts the
migration work.