(MAINT) Assume all env. vars. are scoped for Make
Previously, component-level env. vars. were scoped for the Make rules,
and would be interpolated/rendered by the shell upon execution. Now,
all env. vars. will be interpolated/rendered by Make before rules are
run. This means that we have to catch values like $$PATH and coerce
them to $(PATH), but it also means that there will be a single,
defined style of env. var (and a single logical interpreter) for
Vanagon projects.