(MAINT) Use more descriptive null commands
A number of methods just returned the string ':', which is *fine*, but
there's no reason we can't be a little more descriptive about what
those null operators actually represent.

(For those unaware, any string that follows the null operator is
treated like an argument to a command that does nothing. It's a safe
way to leave a comment directly in code that will otherwise be
executed.)