(RE-5646) add_source unpacks compressed second source & added spec tests
Before, if a second source was compressed it could not be unpacked because we assumed it was along the lines of a conf file. Now handles second sources that are archives and unpacks them appropriately. Spec tests were also added to ensure that the secondary source was added and unpacked properly.