(VANAGON-47) Simplify profiling_shell.sh
Because of the platform-by-platform quirks of sending UDP messages or
converting a decimal to an integer, `profiling_shell.sh` will now write
run times into a batch file for each Make target it executes.

Every status in the ${PIPESTATUS} array will now be checked, in an
attempt to avoid setting `-o pipefail`, which is more strict than Make
or Bash usually would be about failed jobs in a pipeline (say, using
the || operator to short-circuit feeding xargs from either of two
similar-but-different `find` commands).

Simplifying `profiling_shell.sh` has also had the side-effect of
removing the dependency on `bc`.