(maint) move add_build_repository and provision_with to Platform
Previously these methods were only accessible from the platform DSL.
Since they are exclusively operating on variables inside the Platform we
can move the methods to the Platform and update the methods in the DSL
to call the Platform methods instead of setting the variables directly.
This will also allow these methods to be called outside of the platform
definition.