(MAINT) Don't use a method that doesn't exist
This has never worked, and it used to silently fail in the background.
This means we'd often end up with rendered nuget templates that looked
like this:

<packageSourceUrl>https://github.com/puppetlabs/-vanagon</packageSourceUrl>
<projectSourceUrl>https://github.com/puppetlabs/</projectSourceUrl>

The nuget template has been updated to use @homepage, which will at
least define a URL if called. We can think about adding additional
configuration atoms for Windows platforms if needed but this will get
the job done.