(maint) Ensure dirname override will be honored
We allow the user to set dirname in case they are dealing with a source
that unpacks into a directory that's different than the name of the
archive. However, we don't honor this. The code always overwrites that
user-set variable with what it thinks it should be. That's lame.

This commit fixes how that variable is treated. If the user does in fact
supply `dirname` in the component config, then that variable assignment
will be honored. Otherwise, it defaults to use the same name as the
original archive.