(MAINT) Git tests shouldn't hit the network
Since we outsource git repo handling to a 3rd party library, we don't
have to test every single aspect of our wrapper. Just ensuring that it
talks to the underlying Git library should be sufficient.

The benefit is that unit test run time drops from about 7 seconds to
consistently less than 1 second. A win is a win, right?